Move some items to nav bar component, use module nav component as container
