+++ src/components/enterprise-wc.ts
@@ -55,6 +55,7 @@ import './ids-message/ids-message';
 import './ids-modal/ids-modal';
 import './ids-month-view/ids-month-view';
 import './ids-module-nav/ids-module-nav';
+import './ids-module-nav/ids-module-nav-bar';
 import './ids-module-nav/ids-module-nav-content';
 import './ids-multiselect/ids-multiselect';
 import './ids-notification-banner/ids-notification-banner';
