+++ src/components/ids-module-nav/demos/example.ts
@@ -1,17 +1,29 @@
 import '../../ids-accordion/ids-accordion';
+import '../../ids-dropdown/ids-dropdown';
 import '../../ids-header/ids-header';
 import '../../ids-search-field/ids-search-field';
 import '../../ids-toolbar/ids-toolbar';
 import '../../ids-toolbar/ids-toolbar-section';
 
-let menuState = 'collapsed';
+import type { IdsModuleNavDisplayMode } from '../ids-module-nav-common';
+
+let menuState: IdsModuleNavDisplayMode = 'collapsed';
 
 document.addEventListener('DOMContentLoaded', () => {
-  const moduleNavDrawer: any = document.querySelector('#module-nav');
+  const moduleNav: any = document.querySelector('ids-module-nav');
+  const moduleNavDrawer: any = document.querySelector('ids-module-nav-bar');
   const appMenuTriggerBtn: any = document.querySelector('#module-nav-trigger');
+  const displayModeDropdown: any = document.querySelector('#dd-display-mode-setting');
+
   moduleNavDrawer.target = appMenuTriggerBtn;
 
-  // =============================
+  const updateDisplayMode = (val: IdsModuleNavDisplayMode) => {
+    menuState = val === 'collapsed' ? 'expanded' : 'collapsed';
+    moduleNav.displayMode = val;
+    console.info('Module Nav Display Mode Updated:', val || 'hidden');
+  };
+
+  // ============================
   // Setup events
 
   moduleNavDrawer.addEventListener('selected', (e: CustomEvent) => {
@@ -19,12 +31,16 @@ document.addEventListener('DOMContentLoaded', () => {
   });
 
   appMenuTriggerBtn.addEventListener('click', () => {
-    menuState = menuState === 'collapsed' ? 'expanded' : 'collapsed';
-    moduleNavDrawer.displayMode = menuState;
+    updateDisplayMode(menuState);
+  });
+
+  displayModeDropdown.addEventListener('change', (e: CustomEvent) => {
+    const selectedValue = e.detail.value;
+    updateDisplayMode(selectedValue);
   });
 
   // =============================
   // Set initial
 
-  moduleNavDrawer.displayMode = menuState;
+  moduleNav.displayMode = menuState;
 });
