+++ src/components/ids-module-nav/demos/index.ts
@@ -1,1 +1,3 @@
 import '../ids-module-nav';
+import '../ids-module-nav-bar';
+import '../ids-module-nav-content';
