+++ src/components/ids-module-nav/ids-module-nav-content.ts
@@ -1,7 +1,9 @@
 import { attributes } from '../../core/ids-attributes';
 import { customElement, scss } from '../../core/ids-decorators';
+import { getClosest } from '../../utils/ids-dom-utils/ids-dom-utils';
 import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
 import IdsElement from '../../core/ids-element';
+
 import IdsModuleNavDisplayModeMixin from './ids-module-nav-display-mode-mixin';
 
 import styles from './ids-module-nav-content.scss';
@@ -12,12 +14,12 @@ const Base = IdsModuleNavDisplayModeMixin(
 
 /**
  * IDS Module Nav Component
- * @type {IdsModuleNav}
- * @inherits IdsDrawer
+ * @type {IdsModuleNavContent}
+ * @inherits IdsElement
  */
 @customElement('ids-module-nav-content')
 @scss(styles)
-export default class IdsModuleNav extends Base {
+export default class IdsModuleNavContent extends Base {
   constructor() {
     super();
   }
@@ -35,6 +37,16 @@ export default class IdsModuleNav extends Base {
 
   template() {
     return `<div class="ids-module-nav-content"><slot></slot></div>`;
+  }
+
+  /**
+   * @readonly
+   * @returns {IdsModuleNav | undefined} reference to the Module Nav parent element
+   */
+  get parent() {
+    const parentEl = getClosest(this, 'ids-module-nav');
+    if (parentEl) return parentEl;
+    return undefined;
   }
 
   /**
