+++ src/enterprise-wc.ts
@@ -54,6 +54,7 @@ import './components/ids-menu-button/ids-menu-button';
 import './components/ids-message/ids-message';
 import './components/ids-modal/ids-modal';
 import './components/ids-module-nav/ids-module-nav';
+import './components/ids-module-nav/ids-module-nav-bar';
 import './components/ids-module-nav/ids-module-nav-content';
 import './components/ids-month-view/ids-month-view';
 import './components/ids-multiselect/ids-multiselect';
